+++ lib/dart_meteor.dart
@@ -1,6 +1,6 @@
 /// Meteor connection for dart
 /// This library is a meteor client warpper
-library dart_meteor;
+library;
 
 export 'src/meteor_client.dart';
 export 'src/ddp_client.dart';